requirements.txt
# Dependências do projeto MONITOUR Turismo & Viagens

# Core Django
Django>=5.2
Pillow>=10.0.0
python-decouple>=3.8
python-dotenv>=1.0.0

# Banco de dados
psycopg2-binary>=2.9.0
mysqlclient>=2.2.0

# Interface administrativa
django-admin-interface>=0.19.0
django-colorfield>=0.11.0
django-import-export>=4.0.0

# Validação e formulários
django-phonenumber-field[phonenumbers]>=3.0.0
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0

# API REST
djangorestframework>=3.15.0
djangorestframework-api-key>=3.0.0
django-cors-headers>=4.3.0

# Outros utilitários
weasyprint>=60.0
ofxparse>=0.21

# Para produção
dj-database-url>=2.0.0
django-storages>=1.14.0
gunicorn>=21.0.0
whitenoise>=6.5.0

# Desenvolvimento
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
Faker>=20.1.0
requests>=2.31.0